+++ HISTORY.md
@@ -1,3 +1,9 @@
+0.5.0 / 2020-06-04
+==================
+
+  * Add support for reading multiple cookies with same name
+    - This feature can be turned on by `multiValuedCookies` flag
+
 0.4.1 / 2020-04-21
 ==================